+++ R/OCN_to_igraph.R
@@ -7,7 +7,7 @@ OCN_to_igraph <- function (OCN,
   if (!(level %in% c("FD","RN","AG"))) {
     stop("Invalid level")
   }
-  if (!("RN" %in% names(OCN)) && (level %in% c("RN","AG"))){
+  if (length(OCN$RN$nNodes)==0 && (level %in% c("RN","AG"))){
     stop('Missing aggregation level in OCN. Run landscape_OCN and/or aggregate_OCN prior to OCN_to_SSN.')
   }