+++ R/draw_elev2D_OCN.R
@@ -1,7 +1,12 @@
 
 draw_elev2D_OCN <- function(OCN,
-                              colPalette=terrain.colors(1000,alpha=1),
-                              addLegend=TRUE){
+                            colPalette=terrain.colors(1000,alpha=1),
+                            addLegend=TRUE,
+                            drawRiver=FALSE,
+                            thrADraw=0.002*OCN$FD$nNodes*OCN$cellsize^2,
+                            riverColor="#00BFFF",
+                            min_lwd=0.5,
+                            max_lwd=5){
   
   if (!("Z" %in% names(OCN$FD))){
     stop('Missing fields in OCN. You should run landscape_OCN prior to draw_elev2D_OCN.')
@@ -33,11 +38,22 @@ draw_elev2D_OCN <- function(OCN,
   
   if (addLegend==TRUE){
   image.plot(Xvec, Yvec,
-             t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1)
+             t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1,
+             xlim=range(OCN$FD$X), ylim=range(OCN$FD$Y))
   } else {
     image(seq(xllcorner,xllcorner+(OCN$dimX-1)*OCN$cellsize,OCN$cellsize),
           seq(yllcorner,yllcorner+(OCN$dimY-1)*OCN$cellsize,OCN$cellsize),
-               t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1)
+               t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1,
+          xlim=range(OCN$FD$X), ylim=range(OCN$FD$Y))
   }
+  
+  if (drawRiver==TRUE){
+    AvailableNodes <- setdiff(1:OCN$FD$nNodes,OCN$FD$outlet)
+    for (i in AvailableNodes){
+      if (OCN$FD$A[i]>=thrADraw  & abs(OCN$FD$X[i]-OCN$FD$X[OCN$FD$downNode[i]]) <= 1.001*OCN$cellsize & abs(OCN$FD$Y[i]-OCN$FD$Y[OCN$FD$downNode[i]]) <= 1.001*OCN$cellsize) {
+        lines(c(OCN$FD$X[i],OCN$FD$X[OCN$FD$downNode[i]]),c(OCN$FD$Y[i],OCN$FD$Y[OCN$FD$downNode[i]]),
+              lwd=min_lwd+(max_lwd-min_lwd)*(OCN$FD$A[i]/(OCN$FD$nNodes*OCN$cellsize^2))^0.5,col=riverColor)}}
+  }
+  
   invisible()
 }