+++ R/draw_thematic_OCN.R
@@ -9,12 +9,12 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
                               chooseCM=FALSE,
                               drawNodes=FALSE,
                               nodeType="upstream",
-                              cex=2,
-                              pch=21,
                               nanColor="#00BFFF",
                               riverColor="#00BFFF",
                               backgroundColor="#999999",
                               addLegend=TRUE,
+                              min_lwd=0.5,
+                              max_lwd=5,
                               ...){
   
   dots <- list(...)
@@ -24,7 +24,8 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
   if (is.null(dots$ylab)){dots$ylab <- ""}
   if (is.null(dots$type)){dots$type <- "n"}
   if (is.null(dots$asp)){dots$asp <- 1}
-
+  if (is.null(dots$cex)){dots$cex <- 2}; cex <- dots$cex
+  if (is.null(dots$pch)){dots$pch <- 21}; pch <- dots$pch
   
   # initialization
   if (!("RN" %in% names(OCN))){ # try to swap arguments if in wrong order
@@ -33,13 +34,9 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
     theme <- tmp
   }
   
-  if (all(is.na(theme))){
-    addLegend <- FALSE
-  }
+  if (all(is.na(theme))) {addLegend <- FALSE }
   
-  if (!("RN" %in% names(OCN))){
-    stop('Missing fields in OCN. You should run aggregate_OCN prior to draw_thematic_OCN.')
-  }
+  if (length(OCN$RN$nNodes)==0){stop('Missing fields in OCN. You should run aggregate_OCN prior to draw_thematic_OCN.')}
   
   if (discreteLevels == FALSE) {
     if (length(colLevels)<3){N_colLevels <- 1000} else {N_colLevels <- colLevels[3]}
@@ -102,13 +99,13 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
   }
   
   if (exactDraw==TRUE){
-    X <- OCN$FD$XDraw#[which( OCN$FD$toCM %in% chooseCM )]
-    Y <- OCN$FD$YDraw#[which( OCN$FD$toCM %in% chooseCM )]
+    X <- OCN$FD$XDraw[which(OCN$FD$toRN!=0)] # at RN level
+    Y <- OCN$FD$YDraw[which(OCN$FD$toRN!=0)] # at RN level
     Xc <- OCN$CM$XContourDraw
     Yc <- OCN$CM$YContourDraw
   } else {
-    X <- OCN$FD$X#[which( OCN$FD$toCM %in% chooseCM )]
-    Y <- OCN$FD$Y#[which( OCN$FD$toCM %in% chooseCM )]
+    X <- OCN$RN$X
+    Y <- OCN$RN$Y
     Xc <- OCN$CM$XContour
     Yc <- OCN$CM$YContour
   }
@@ -121,10 +118,10 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
     pch_vec <- pch*rep(1,length(theme))
   } else {pch_vec <- pch}
   
-  AvailableNodes <- setdiff(which(OCN$FD$toCM %in% chooseCM),OCN$FD$outlet)
+  AvailableNodes <- setdiff(which(OCN$RN$toCM %in% chooseCM),OCN$RN$outlet)
 
-  if (is.null(dots$xlim)){ dots$xlim <- c(min(X[OCN$FD$toCM %in% chooseCM]),max(X[OCN$FD$toCM %in% chooseCM])) }
-  if (is.null(dots$ylim)){ dots$ylim <- c(min(Y[OCN$FD$toCM %in% chooseCM]),max(Y[OCN$FD$toCM %in% chooseCM])) }
+  if (is.null(dots$xlim)){ dots$xlim <- c(min(X[OCN$RN$toCM %in% chooseCM]),max(X[OCN$RN$toCM %in% chooseCM])) }
+  if (is.null(dots$ylim)){ dots$ylim <- c(min(Y[OCN$RN$toCM %in% chooseCM]),max(Y[OCN$RN$toCM %in% chooseCM])) }
   dots$x <- dots$xlim
   dots$y <- dots$ylim
   
@@ -149,22 +146,22 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
   }
   
   for (i in AvailableNodes){
-    rn <- OCN$FD$toRN[i]
-    reach <- OCN$RN$toAGReach[rn]
-    if (OCN$FD$A[i]>=OCN$thrA & 
-        abs(X[i]-X[OCN$FD$downNode[i]]) <= 1.001*OCN$cellsize & 
-        abs(Y[i]-Y[OCN$FD$downNode[i]]) <= 1.001*OCN$cellsize &
+    #rn <- OCN$FD$toRN[i]
+    reach <- OCN$RN$toAGReach[i]
+    if (OCN$RN$A[i]>=OCN$thrA & 
+        abs(X[i]-X[OCN$RN$downNode[i]]) <= 1.001*OCN$cellsize & 
+        abs(Y[i]-Y[OCN$RN$downNode[i]]) <= 1.001*OCN$cellsize &
         X[i] >= xy_lim[1]-1*OCN$cellsize &
         X[i] <= xy_lim[2]+1*OCN$cellsize &
         Y[i] >= xy_lim[3]-1*OCN$cellsize & 
         Y[i] <= xy_lim[4]+1*OCN$cellsize) {
-      if ( (byRN==TRUE && (is.nan(theme[rn])==TRUE | is.na(theme[rn])==TRUE)) ||
+      if ( (byRN==TRUE && (is.nan(theme[i])==TRUE | is.na(theme[i])==TRUE)) ||
            (byRN==FALSE && (is.nan(theme[reach])==TRUE | is.na(theme[reach])==TRUE)) ||
-           (byRN==TRUE && cutoff==TRUE && (theme[rn] < min(Breakpoints) || theme[rn] > max(Breakpoints))) ||
+           (byRN==TRUE && cutoff==TRUE && (theme[i] < min(Breakpoints) || theme[i] > max(Breakpoints))) ||
            (byRN==FALSE && cutoff==TRUE && (theme[reach] < min(Breakpoints) || theme[reach] > max(Breakpoints))) )  {
         hexcolor <- nanColor
       } else {
-        if (byRN==TRUE){val <- theme[rn]} else {val <- theme[reach]}
+        if (byRN==TRUE){val <- theme[i]} else {val <- theme[reach]}
         
         colvalue <- which(Breakpoints > val)[1] - 1  
         if (isTRUE(colvalue==0)) {colvalue <- 1}
@@ -173,11 +170,11 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
         hexcolor <- colPalette[colvalue]
       }
       if (drawNodes==TRUE){
-        lines(c(X[i],X[OCN$FD$downNode[i]]),c(Y[i],Y[OCN$FD$downNode[i]]),
-              lwd=0.5+4.5*(OCN$FD$A[i]/max(OCN$FD$A[AvailableNodes]))^0.5,col=riverColor)
+        lines(c(X[i],X[OCN$RN$downNode[i]]),c(Y[i],Y[OCN$RN$downNode[i]]),
+              lwd=min_lwd+(max_lwd-min_lwd)*(OCN$RN$A[i]/max(OCN$RN$A[AvailableNodes]))^0.5,col=riverColor)
       } else {
-        lines(c(X[i],X[OCN$FD$downNode[i]]),c(Y[i],Y[OCN$FD$downNode[i]]),
-              lwd=0.5+4.5*(OCN$FD$A[i]/max(OCN$FD$A[AvailableNodes]))^0.5,col=hexcolor)}
+        lines(c(X[i],X[OCN$RN$downNode[i]]),c(Y[i],Y[OCN$RN$downNode[i]]),
+              lwd=min_lwd+(max_lwd-min_lwd)*(OCN$RN$A[i]/max(OCN$RN$A[AvailableNodes]))^0.5,col=hexcolor)}
     }
   }
   
@@ -220,11 +217,11 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
         hexcolor <- colPalette[colvalue]
       }
       if (byRN==TRUE){
-        node <- which(OCN$FD$X==OCN$RN$X[i] & OCN$FD$Y==OCN$RN$Y[i])
+        node <- 1:OCN$RN$nNodes #which(OCN$FD$X==OCN$RN$X[i] & OCN$FD$Y==OCN$RN$Y[i])
       } else if (byRN==FALSE && nodeType=="upstream") {
-        node <- which(OCN$FD$X==OCN$AG$X[i] & OCN$FD$Y==OCN$AG$Y[i])
+        node <- which(OCN$RN$X==OCN$AG$X[i] & OCN$RN$Y==OCN$AG$Y[i])
       } else if (byRN==FALSE && nodeType=="downstream") {
-        node <- which(OCN$FD$X==OCN$AG$XReach[i] & OCN$FD$Y==OCN$AG$YReach[i])
+        node <- which(OCN$RN$X==OCN$AG$XReach[i] & OCN$RN$Y==OCN$AG$YReach[i])
       }
       points(X[node],Y[node],bg=hexcolor,pch=pch_vec[i],cex=cex_vec[i])  
     }
@@ -252,7 +249,7 @@ draw_thematic_OCN <- function(OCN,theme=NA*numeric(OCN$AG$nNodes),
                                   as.character(round(1000*Breakpoints[N_colLevels+1])/1000),"]",sep="")
       }
       
-      legend(x=1.01*max(X[OCN$FD$toCM %in% chooseCM]),y=max(Y[OCN$FD$toCM %in% chooseCM]),
+      legend(x=1.01*max(X[OCN$RN$toCM %in% chooseCM]),y=max(Y[OCN$RN$toCM %in% chooseCM]),
              str,fill=colPalette,ncol=ceiling(N_colLevels/20), xpd=TRUE, cex=0.8, bty="n")
     }
   }